--- src/SZahir/Week01/OddEven.java
@@ -1,18 +1,0 @@
-package SZahir.Week01;
-
-public class OddEven {
-    public static void main(String[] args) {
-        oddEven(9);
-
-
-    }
-    public static int oddEven(int num1){
-        if(!( num1 %2==0)) {
-            System.out.println(num1 + " is odd");
-        } else {
-            System.out.println(num1 + " is even");
-        }
-        return num1;
-    }
-}
-
