--- src/SZahir/Week04/Tasks.java
@@ -1,21 +1,0 @@
-package SZahir.Week04;
-
-public class Tasks {
-    /*
-    String -- Same letters
-
-1. Write a return method that checks if a string is built out of the same letters as another string.
-Ex:  same("abc",  "cab"); -> true
-same("abc",  "abb"); -> false
-
-2. String -- Remove Duplicates
-
-Write a return method that can remove the duplicated values from String
-Ex:  removeDup("AAABBBCCC")  ==> ABC
-
-3.String -- Find the unique
-Write a return  method that can find the unique characters from the String
-Ex:  unique("AAABBBCCCDEF")  ==>  "DEF";
-
-     */
-}
